+++ usaspending_api/etl/transaction_loaders/derived_field_functions_fpds.py
@@ -1,7 +1,7 @@
 from datetime import datetime, timezone
 
 from usaspending_api.common.helpers.date_helper import fy
-from usaspending_api.data_load.cached_reference_data import subtier_agency_list
+from usaspending_api.etl.transaction_loaders.cached_reference_data import subtier_agency_list
 from usaspending_api.broker.helpers.get_business_categories import get_business_categories
