+++ usaspending_api/etl/transaction_loaders/field_mappings_fpds.py
@@ -1,4 +1,4 @@
-from usaspending_api.data_load.derived_field_functions_fpds import (
+from usaspending_api.etl.transaction_loaders.derived_field_functions_fpds import (
     calculate_fiscal_year,
     calculate_awarding_agency,
     calculate_funding_agency,
@@ -6,7 +6,7 @@ from usaspending_api.data_load.derived_field_functions_fpds import (
     current_datetime,
     business_categories,
 )
-from usaspending_api.data_load.data_load_helpers import capitalize_if_string
+from usaspending_api.etl.transaction_loaders.data_load_helpers import capitalize_if_string
 
 # broker column name -> usaspending column name
 transaction_fpds_columns = {
