+++ usaspending_api/etl/transaction_loaders/tests/test_data_load_helpers.py
@@ -1,4 +1,4 @@
-from usaspending_api.data_load.data_load_helpers import capitalize_if_string, false_if_null
+from usaspending_api.etl.transaction_loaders.data_load_helpers import capitalize_if_string, false_if_null
 
 
 def test_capitalize_if_string():
